moves up state from book table to cube component
